pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tibcleaner"
version = "0.0.1"
authors = [
  { name="OpenPecha", email="dev@openpecha.org" },
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "beautifulsoup4>=4.12.3",
  "EbookLib>=0.18",
  "exceptiongroup>=1.2.0",
  "lxml>=5.1.0",
  "pdf2image>=1.17.0",
  "Pillow>=10.2.0",
  "pypdf>=4.1.0",
  "python-docx>=1.1.0",
  "PyYAML>=6.0.1",
  "rarfile>=4.1",
  "striprtf>=0.0.26",
  "tqdm>=4.66.2",
  "botok>=0.8.12",
  "chardet>=5.2.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pre-commit",
]


[project.urls]
"Homepage" = "https://github.com/OpenPecha/openpecha-project-template"
"Bug Tracker" = "https://github.com/OpenPecha/openpecha-project-template/issues"
